Add polyfill to localStorage

+++ src/lib/tubular-settings.service.ts
@@ -9,17 +9,29 @@ export interface ITubularSettingsProvider {
 }
 
 export class TubularLocalStorageService implements ITubularSettingsProvider {
-    // TODO: Add Check if localStorage exists at constructor
+
+    existLocalStorage: boolean = true;
+    _data: {};
+
+    constructor() {
+        if (!window.localStorage) {
+            this.existLocalStorage = false;
+            console.log("Browser does not support localStorage");
+        }
+    }
     
     public put(id: string, value: string) {
-        localStorage.setItem(id, JSON.stringify(value));
+        if (this.existLocalStorage) localStorage.setItem(id, JSON.stringify(value));
+        else this._data[id] = String(value);
     }
 
     public get(key: string): any {
-        return JSON.parse(localStorage.getItem(key)) || false;
+        if (this.existLocalStorage) return JSON.parse(localStorage.getItem(key)) || false;
+        else return this._data.hasOwnProperty(key) ? this._data[key] : false;
     }
     
     public delete(key: string) {
-        localStorage.removeItem(key);
+        if (this.existLocalStorage) localStorage.removeItem(key);
+        else return delete this._data[key];
     }
 }
